backend: (riscv) flatten loops that don't use induction variables
