+++ tests/filecheck/dialects/riscv_scf/loop_flatten.mlir
@@ -1,12 +1,21 @@
-// RUN: xdsl-opt -p riscv-scf-loop-fusion %s | filecheck %s
+// RUN: xdsl-opt -p riscv-scf-loop-flatten %s | filecheck %s
 
 // CHECK:       builtin.module {
 
-// Success case
+// Success cases
 %c0 = riscv.li 0 : () -> !riscv.reg<>
 %c1 = riscv.li 1 : () -> !riscv.reg<>
+%c3 = riscv.li 3 : () -> !riscv.reg<>
+%c5 = riscv.li 5 : () -> !riscv.reg<>
 %c8 = riscv.li 8 : () -> !riscv.reg<>
 %c64 = riscv.li 64 : () -> !riscv.reg<>
+
+// CHECK-NEXT:    %c0 = riscv.li 0 : () -> !riscv.reg<>
+// CHECK-NEXT:    %c1 = riscv.li 1 : () -> !riscv.reg<>
+// CHECK-NEXT:    %c3 = riscv.li 3 : () -> !riscv.reg<>
+// CHECK-NEXT:    %c5 = riscv.li 5 : () -> !riscv.reg<>
+// CHECK-NEXT:    %c8 = riscv.li 8 : () -> !riscv.reg<>
+// CHECK-NEXT:    %c64 = riscv.li 64 : () -> !riscv.reg<>
 
 riscv_scf.for %16 : !riscv.reg<> = %c0 to %c64 step %c8 {
     riscv_scf.for %17 : !riscv.reg<> = %c0 to %c8 step %c1 {
@@ -16,14 +25,26 @@ riscv_scf.for %16 : !riscv.reg<> = %c0 to %c64 step %c8 {
     }
 }
 
-// CHECK-NEXT:    %c0 = riscv.li 0 : () -> !riscv.reg<>
-// CHECK-NEXT:    %c1 = riscv.li 1 : () -> !riscv.reg<>
-// CHECK-NEXT:    %c8 = riscv.li 8 : () -> !riscv.reg<>
-// CHECK-NEXT:    %c64 = riscv.li 64 : () -> !riscv.reg<>
 // CHECK-NEXT:    riscv_scf.for %0 : !riscv.reg<> = %c0 to %c64 step %c1 {
 // CHECK-NEXT:      %1 = riscv.li 8 : () -> !riscv.reg<>
 // CHECK-NEXT:      "test.op"(%0) : (!riscv.reg<>) -> ()
 // CHECK-NEXT:    }
+
+riscv_scf.for %i : !riscv.reg<> = %c0 to %c64 step %c5 {
+    riscv_scf.for %j : !riscv.reg<> = %c0 to %c8 step %c3 {
+        %k = riscv.li 8 : () -> !riscv.reg<>
+        "test.op"(%k) : (!riscv.reg<>) -> ()
+    }
+}
+
+// CHECK-NEXT:    %{{.*}} = riscv.li 2 : () -> !riscv.reg<>
+// CHECK-NEXT:    %{{.*}} = riscv.mul %c64, %{{.*}} : (!riscv.reg<>, !riscv.reg<>) -> !riscv.reg<>
+// CHECK-NEXT:    riscv_scf.for %{{.*}} : !riscv.reg<> = %c0 to %{{.*}} step %c5 {
+// CHECK-NEXT:      %{{.*}} = riscv.li 8 : () -> !riscv.reg<>
+// CHECK-NEXT:      "test.op"(%{{.*}}) : (!riscv.reg<>) -> ()
+// CHECK-NEXT:    }
+
+// Failures add induction variables:
 
 // Cannot fuse outer loop with iteration arguments
 %res0 = riscv_scf.for %16 : !riscv.reg<> = %c0 to %c64 step %c8 iter_args(%arg0 = %c0) -> (!riscv.reg<>) {
@@ -145,6 +166,37 @@ riscv_scf.for %16 : !riscv.reg<> = %c0 to %c64 step %c8 {
 // CHECK-NEXT:      }
 // CHECK-NEXT:    }
 
+// Failures no induction variables:
+
+riscv_scf.for %i : !riscv.reg<> = %c1 to %c64 step %c5 {
+    riscv_scf.for %j : !riscv.reg<> = %c0 to %c8 step %c3 {
+        %k = riscv.li 8 : () -> !riscv.reg<>
+        "test.op"(%k) : (!riscv.reg<>) -> ()
+    }
+}
+
+// CHECK-NEXT:    riscv_scf.for %{{.*}} : !riscv.reg<> = %c1 to %c64 step %c5 {
+// CHECK-NEXT:        riscv_scf.for %{{.*}} : !riscv.reg<> = %c0 to %c8 step %c3 {
+// CHECK-NEXT:            %{{.*}} = riscv.li 8 : () -> !riscv.reg<>
+// CHECK-NEXT:            "test.op"(%{{.*}}) : (!riscv.reg<>) -> ()
+// CHECK-NEXT:        }
+// CHECK-NEXT:    }
+
+%non_const = "test.op"() : () -> !riscv.reg<>
+// CHECK-NEXT:    %non_const = "test.op"() : () -> !riscv.reg<>
+
+riscv_scf.for %i : !riscv.reg<> = %non_const to %c64 step %c5 {
+    riscv_scf.for %j : !riscv.reg<> = %c0 to %c8 step %c3 {
+        %k = riscv.li 8 : () -> !riscv.reg<>
+        "test.op"(%k) : (!riscv.reg<>) -> ()
+    }
+}
+
+// CHECK-NEXT:    riscv_scf.for %{{.*}} : !riscv.reg<> = %non_const to %c64 step %c5 {
+// CHECK-NEXT:        riscv_scf.for %{{.*}} : !riscv.reg<> = %c0 to %c8 step %c3 {
+// CHECK-NEXT:            %{{.*}} = riscv.li 8 : () -> !riscv.reg<>
+// CHECK-NEXT:            "test.op"(%{{.*}}) : (!riscv.reg<>) -> ()
+// CHECK-NEXT:        }
+// CHECK-NEXT:    }
+
 // CHECK-NEXT:  }
-
-
